% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGE.R
\name{plotGE}
\alias{plotGE}
\title{plotGE}
\usage{
plotGE(
  spe,
  gene_id = NULL,
  gene_name = NULL,
  log_count = FALSE,
  Visium = FALSE,
  ...
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data object.}

\item{gene_id}{(character) The id of one gene. Refer to gene_id Column in \code{rowData}}

\item{gene_name}{(character) The name of one gene. Refer to gene_name Column in \code{rowData}}

\item{log_count}{(logical) Whether to plot the log transfered counts if exists. Default
= FALSE.}

\item{Visium}{(logical) Whether to show histology image as background as in \link{plotVisium}. Default
= FALSE.}

\item{...}{Other arguments for \link{plotVisium} and \link{plotSpots}}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, customized formatting, etc).
}
\description{
Plot gene expression for spatially resolved transcriptomics data
}
\details{
Function to generate gene expression plots for spatially resolved transcriptomics datasets

This function generates a gene expression plot for spot-based spatially resolved
transcriptomics data with several
options available to adjust the plot type and style.
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
plotGE(spe, gene_id = "ENSG00000243485", Visium = TRUE) 
# Equivalently 
plotGE(spe, gene_name = "MIR1302-2HG", Visium = TRUE) 
}
