% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{plotQC}
\usage{
plotQC(
  spe,
  type = c("bar", "scatter", "spots"),
  x_coord = "x",
  y_coord = "y",
  in_tissue = TRUE,
  metric_x = "cell_count",
  metric_y = "sum",
  discard = "discard",
  highlight_zeros = TRUE,
  threshold_x = NULL,
  threshold_y = NULL,
  trend = TRUE,
  marginal = TRUE,
  y_reverse = TRUE
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
`SpatialExperiment` object.}

\item{type}{(character) Type of QC plot. Options are `bar`, `scatter`, and
`spots`. See details in description.}

\item{x_coord}{(character) Name of column in `spatialCoords` containing
x-coordinates. Default = "x". Required for spot-based plots.}

\item{y_coord}{(character) Name of column in `spatialCoords` containing
y-coordinates. Default = "y". Required for spot-based plots.}

\item{in_tissue}{(logical) Whether to show only spots over tissue, or all
spots. Options are `TRUE` (show spots over tissue; requires a column
labelled `in_tissue` in `spatialData` identifying spots over tissue, which
is the standard format for 10x Genomics Visium data), `FALSE` (show all
spots), or a character value with the name of a column in `spatialData`
identifying the spots to show.}

\item{metric_x}{(character) Name of column in `colData` containing QC metric
to plot on x-axis (e.g. "cell_count" for number of cells per spot). Default
= "cell_count". Required for barplots and scatterplots.}

\item{metric_y}{(character) Name of column in `colData` containing QC metric
to plot on y-axis (e.g. "sum" for number of detected transcripts, or
"detected" for number of detected genes). Default = "sum". Required for
scatterplots.}

\item{discard}{(character) Name of column in `colData` identifying discarded
spots that do not meet filtering thresholds, which will be highlighted on a
spot-based plot. Default = "discard". Optional for spot-based plots.}

\item{highlight_zeros}{(logical) Whether to highlight bar for x = 0 (e.g.
zero cells per spot). Default = TRUE. Optional for barplots.}

\item{threshold_x}{(numeric) Filtering threshold for QC metric on x-axis,
which will be highlighted with a vertical bar. Default = NULL. Optional for
scatterplots.}

\item{threshold_y}{(numeric) Filtering threshold for QC metric on y-axis,
which will be highlighted with a horizontal bar. Default = NULL. Optional
for scatterplots.}

\item{trend}{(logical) Whether to display a smoothed trend (loess) for
scatterplots. Default = TRUE. Optional for scatterplots.}

\item{marginal}{(logical) Whether to display marginal histograms for
scatterplots. Default = TRUE. Optional for scatterplots.}

\item{y_reverse}{(logical) Whether to reverse `y` coordinates, which is often
required for 10x Genomics Visium data. Default = `TRUE`.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Quality control (QC) plots for spatially resolved transcriptomics data.
}
\details{
Function to generate plots for quality control (QC) purposes for spatially
resolved transcriptomics data.

The following types of QC plots are available:

- Barplot (`type = bar`) for a single QC metric, e.g. number of cells per
spot. For number of cells per spot, the barplot highlights spots with zero
cells.
- Scatterplot (`type = scatter`) comparing two QC metrics, e.g. number of
detected features vs. number of cells per spot, with optional vertical and
horizontal lines highlighting QC filtering thresholds.
- Spots (`type = spots`) i.e. spots in spatial (x-y) coordinates,
highlighting discarded spots that do not meet filtering thresholds.
}
\examples{
# to do

}
