% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQCspots.R
\name{plotQCspots}
\alias{plotQCspots}
\title{plotQCspots}
\usage{
plotQCspots(spe, x_coord = "x", y_coord = "y", discard = "discard")
}
\arguments{
\item{spe}{Input object (SpatialExperiment).}

\item{x_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "x".}

\item{y_coord}{Name of column in spatialCoords slot containing x-coordinates.
Default = "y".}

\item{discard}{Name of column in colData identifying spots to be discarded
(TRUE/FALSE values). Default = "discard".}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Quality control (QC) plots for spatial transcriptomics datasets.
}
\details{
Functions to generate quality control (QC) plots for spatial transcriptomics
datasets.

This function generates a plot identifying spatial coordinates (spots) that
do not meet quality control (QC) metrics (i.e. discarded spots). Spots are
shown in the physical x-y coordinates of the tissue slide. If discarded spots
correspond to known anatomical/histological or other biological regions of
interest, this may be problematic for downstream analyses.
}
\examples{
# to do

}
