% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{plotDimRed}
\usage{
plotDimRed(
  spe,
  type = "UMAP",
  x_axis = "UMAP1",
  y_axis = "UMAP2",
  discrete = NULL,
  continuous = NULL,
  palette = NULL
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data object.}

\item{type}{(character) Type of reduced dimension plot. Options are "PCA" or
"UMAP". Default = "UMAP.}

\item{x_axis}{(character) Name of column in reducedDim slot containing
x-coordinates. Default = "UMAP1".}

\item{y_axis}{(character) Name of column in reducedDim slot containing
y-coordinates. Default = "UMAP2".}

\item{discrete}{(character) Name of column in colData containing discrete
labels (e.g. cluster labels or ground truth labels) to show with colors.
Default = NULL.}

\item{continuous}{(character) Name of column in colData containing continuous
values (e.g. total UMI counts) to show with colors. Default = NULL.}

\item{palette}{(character) Color palette. Options for discrete labels are
"libd_layer_colors", "Okabe-Ito", or a vector of hex codes for a custom
palette. Default = "libd_layer_colors". Options for continuous values are
"navy", or a vector of length two containing custom colors. Default =
"navy".}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, customized formatting, etc).
}
\description{
Plots for spatial transcriptomics datasets.
}
\details{
Function to plot spatial transcriptomics data in reduced dimension
coordinates.

This function generates a plot showing spatial coordinates (spots) in reduced
dimension coordinates (PCA or UMAP), with optional colors for cluster labels,
ground truth labels, or other quantities.
}
\examples{
# to do

}
