% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{plotDimRed}
\usage{
plotDimRed(
  spe,
  type = "UMAP",
  x_axis = "UMAP1",
  y_axis = "UMAP2",
  cluster_id = NULL,
  palette = "libd_layer_colors"
)
}
\arguments{
\item{spe}{Input object (SpatialExperiment).}

\item{type}{Type of dimension reduction to use. Options are "PCA" or "UMAP".
Default = "UMAP".}

\item{x_axis}{Name of column in reducedDim to use for x-axis. Default =
"UMAP1".}

\item{y_axis}{Name of column in reducedDim to use for y-axis. Default =
"UMAP2".}

\item{cluster_id}{Name of column in colData containing cluster IDs. To plot
without cluster labels, set "cluster_id = NULL". Default = "cluster_id".}

\item{palette}{Color palette for cluster labels. Options are
"libd_layer_colors", "Okabe-Ito", or providing a vector of hex codes for a
custom palette.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, formatting).
}
\description{
Plots for spatial transcriptomics datasets.
}
\details{
Functions to generate plots of spatial coordinates for spatial
transcriptomics datasets, optionally including cluster labels.

This function generates a plot showing spatial coordinates (spots) in reduced
dimension space (either PCA or UMAP). Cluster labels or ground truth labels
can be shown with colors.
}
\examples{
# to do

}
