% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureQC.R
\name{plotFeatureQC}
\alias{plotFeatureQC}
\title{plotFeatureQC}
\usage{
plotFeatureQC(
  spe,
  type = c("hist", "violin"),
  metric_x = "sum",
  annotate = NULL,
  nbins = 100,
  pt.size = 0.1,
  marginal = TRUE,
  y_reverse = TRUE
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
\code{SpatialExperiment} or \code{SingleCellExperiment} object.}

\item{type}{(character) Type of QC plot. Options are "hist", "scatter", and
"spots". See details in description.}

\item{metric_x}{(character) Name of column in \code{colData} containing QC
metric to plot on x-axis (e.g. "sum" for number of cells per spot).
Default = "sum". Required for histogram, scatter, and violin plots.}

\item{annotate}{(logical) Name of column in \code{colData} identifying
flagged spots that do not meet filtering thresholds, which will be
highlighted on a histogram, spot-based, or violin plot. Default = NULL. 
Not needed for scatter plot.}

\item{nbins}{(numeric) Adjusting the histogram bin width. Optional for
spot-based and scatter and violin plots.}

\item{pt.size}{(numeric) Adjusting the scatter, spot-based, violin jitter point size. 
Optional for histogram. Suggested point size for scatter = 0.5, for spot-based
= 0.3, and for violin jitter = 0.1.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Quality control (QC) plots for spatially resolved transcriptomics data.
}
\details{
Function to generate plots for quality control (QC) purposes for spatially
resolved transcriptomics data.

The following types of QC plots are available for per feature QC:

- Histogram (\code{type} = "hist") for a single QC metric, e.g. number of counts
per feature. For number of counts per spot, the histogram highlights spots with
derived flags, e.g. low abundance genes.
- Violin (\code{type} = "violin") for a single QC metric, e.g. number of counts
per feature. For number of counts per spot, the violin plot is able to highlights 
spots with derived flags, e.g. low abundance genes.
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
plotFeatureQC(spe, type = "hist", metric_x = "cell_count", annotate = "low_abun_gene")
plotFeatureQC(spe, type = "scatter", metric_x = "cell_count", metric_y = "sum")
}
