% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMolecules.R
\name{plotMolecules}
\alias{plotMolecules}
\title{plotMolecules}
\usage{
plotMolecules(
  spe,
  molecule = NULL,
  x_coord = "x",
  y_coord = "y",
  palette = c("gray90", "navy"),
  size = 0.3
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
\code{SpatialExperiment} object.}

\item{molecule}{(character) Name of mRNA molecule to plot (assumed to match
one of the row names of \code{rowData}).}

\item{x_coord}{(character) Name of column in \code{spatialCoords} containing
x-coordinates of the cell centroids. Default = "x".}

\item{y_coord}{(character) Name of column in \code{spatialCoords} containing
y-coordinates of the cell centroids. Default = "y".}

\item{palette}{(character) Color palette, provided as a vector of length 2
for the low and high range. Default = \code{c("gray90", "navy")}.}

\item{size}{(numeric) Point size for \code{geom_point()}. Default = 0.3.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Plotting functions for spatially resolved transcriptomics data.
}
\details{
Function to plot molecule-based spatially resolved transcriptomics data
stored in a \code{SpatialExperiment} object.

This function generates a plot in spatial coordinates (e.g. x-y coordinates
on a tissue slide), for a selected molecule.
}
\examples{
library(STexampleData)
spe <- seqFISH_mouseEmbryo()
plotMolecules(spe, molecule = "Sox2")

}
