% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpotQC.R
\name{plotSpotQC}
\alias{plotSpotQC}
\title{plotSpotQC}
\usage{
plotSpotQC(
  spe,
  type = c("hist", "scatter", "spots", "violin"),
  x_coord = NULL,
  y_coord = NULL,
  in_tissue = "in_tissue",
  metric_x = "sum",
  metric_y = "cell_count",
  annotate = NULL,
  nbins = 100,
  pt.size = 0.3,
  threshold_x = NULL,
  threshold_y = NULL,
  trend = TRUE,
  marginal = TRUE,
  y_reverse = TRUE
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
\code{SpatialExperiment} or \code{SingleCellExperiment} object.}

\item{type}{(character) Type of QC plot. Options are "hist", "scatter",
"spots", and "violin". See details in description.}

\item{x_coord}{(character) Name of column in \code{spatialCoords} containing
x-coordinates. Default = NULL, which selects the first column. Used for
spot-based plots.}

\item{y_coord}{(character) Name of column in \code{spatialCoords} containing
y-coordinates. Default = NULL, which selects the second column. Used for
spot-based plots.}

\item{in_tissue}{(character) Name of column in \code{colData} identifying
spots over tissue, e.g. "in_tissue" for 10x Genomics Visium data. If this
argument is provided, only spots over tissue will be shown. Alternatively,
set to NULL to display all spots for Xenium, CosMx, etc. Default = "in_tissue".}

\item{metric_x}{(character) Name of column in \code{colData} containing QC
metric to plot on x-axis (e.g. "sum" for number of cells per spot).
Default = "sum". Required for histogram, scatter, and violin plots.}

\item{metric_y}{(character) Name of column in \code{colData} containing QC
metric to plot on y-axis (e.g. "cell_count" for number of detected transcripts, or
"high_mito" for binary ). Default = "cell_count". Only required for scatter plots.}

\item{annotate}{(logical) Name of column in \code{colData} identifying
flagged spots that do not meet filtering thresholds, which will be
highlighted on a histogram, spot-based, or violin plot. Default = NULL. 
Not needed for scatter plot.}

\item{nbins}{(numeric) Adjusting the histogram bin width. Optional for
spot-based and scatter and violin plots.}

\item{pt.size}{(numeric) Adjusting the scatter, spot-based, violin jitter point size. 
Optional for histogram. Suggested point size for scatter = 0.5, for spot-based
= 0.3, and for violin jitter = 0.1.}

\item{threshold_x}{(numeric) Filtering threshold for QC metric on x-axis,
which will be highlighted with a vertical bar. Default = NULL. Optional for
scatterplots.}

\item{threshold_y}{(numeric) Filtering threshold for QC metric on y-axis,
which will be highlighted with a horizontal bar. Default = NULL. Optional
for scatterplots.}

\item{trend}{(logical) Whether to display a smoothed trend (loess) for
scatterplots. Default = TRUE. Optional for scatterplots.}

\item{marginal}{(logical) Whether to display marginal histograms for
scatterplots. Default = TRUE. Optional for scatterplots.}

\item{y_reverse}{(logical) Whether to reverse y coordinates, which is often
required for 10x Genomics Visium data. Default = TRUE.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Quality control (QC) plots for spatially resolved transcriptomics data.
}
\details{
Function to generate plots for quality control (QC) purposes for spatially
resolved transcriptomics data.

The following types of QC plots are available for per spot/cell QC:

- Histogram (\code{type} = "hist") for a single QC metric, e.g. number of counts
per spot. For number of counts per spot, the histogram highlights spots with
derived flags, e.g. low library size.
- Scatterplot (\code{type} = "scatter") comparing two QC metrics, e.g. number
of detected features vs. number of cells per spot, with optional vertical and
horizontal lines highlighting QC filtering thresholds.
- Spots (\code{type} = "spots") i.e. spots in spatial (x-y) coordinates,
highlighting flagged spots that do not meet filtering thresholds.
- Violin (\code{type} = "violin") for a single QC metric, e.g. number of counts
per spot. For number of counts per spot, the violin plot is able to highlights 
spots with derived flags, e.g. low library size.
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
spe$sum <- colSums(counts(spe))
spe$low_libsize <- spe$sum < 400

plotSpotQC(spe, type = "hist", metric_x = "sum", annotate = "low_libsize")
plotSpotQC(spe, type = "scatter", metric_x = "sum", metric_y = "cell_count")
plotSpotQC(spe, type = "spots", annotate = "low_libsize")
plotSpotQC(spe, type = "violin", metric_x = "sum", annotate = "low_libsize")
}
\author{
Lukas M. Weber with modifications by Yixing E. Dong
}
