% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpots.R
\name{plotSpots}
\alias{plotSpots}
\title{plotSpots}
\usage{
plotSpots(
  spe,
  x_coord = NULL,
  y_coord = NULL,
  sample_id = "sample_id",
  in_tissue = "in_tissue",
  annotate = NULL,
  palette = "libd_layer_colors",
  y_reverse = TRUE,
  size = 0.3
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
\code{SpatialExperiment} object.}

\item{x_coord}{(character) Name of column in \code{spatialCoords} containing
x-coordinates. Default = NULL, which selects the first column.}

\item{y_coord}{(character) Name of column in \code{spatialCoords} containing
y-coordinates. Default = NULL, which selects the second column.}

\item{sample_id}{(character) Name of column in \code{colData} containing
sample IDs. For datasets with multiple samples, this is used to plot
multiple panels (one per sample) using facetting.}

\item{in_tissue}{(character) Name of column in \code{colData} identifying
spots over tissue, e.g. "in_tissue" for 10x Genomics Visium data. If this
argument is provided, only spots over tissue will be shown. Alternatively,
set to NULL to display all spots. Default = "in_tissue".}

\item{annotate}{(character) Name of column in \code{colData} containing
values to annotate spots with colors, e.g. cluster labels (discrete values)
or total UMI counts (continuous values).}

\item{palette}{(character) Color palette for annotation. Options for discrete
labels (e.g. cluster labels) are "libd_layer_colors", "Okabe-Ito", or a
vector of color names or hex values. For continuous values (e.g. total UMI
counts), provide a vector of length 2 for the low and high range, e.g.
\code{c("gray90", "navy")}. Default = \code{"libd_layer_colors"}.}

\item{y_reverse}{(logical) Whether to reverse y coordinates, which is often
required for 10x Genomics Visium data. Default = TRUE.}

\item{size}{(numeric) Point size for \code{geom_point()}. Default = 0.3.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Plotting functions for spatially resolved transcriptomics data.
}
\details{
Function to plot spot-based spatially resolved transcriptomics data stored in
a \code{SpatialExperiment} object.

This function generates a plot in spatial coordinates (e.g. x-y coordinates
on a tissue slide), along with annotation such as cluster labels or total UMI
counts.
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
plotSpots(spe, annotate = "ground_truth")

}
