% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpots.R
\name{plotSpots}
\alias{plotSpots}
\title{plotSpots}
\usage{
plotSpots(
  spe,
  x_coord = NULL,
  y_coord = NULL,
  sample_id = NULL,
  in_tissue = "in_tissue",
  annotate = NULL,
  palette = NULL,
  trans = "identity",
  assay = "counts",
  legend_position = "right",
  point_size = 0.3,
  show_axis = FALSE,
  text_by = NULL,
  text_by_size = 5,
  text_by_color = "black",
  y_reverse = TRUE
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data, assumed to be a
\code{SpatialExperiment} or \code{SingleCellExperiment} object.}

\item{x_coord}{(character) Name of column in \code{spatialCoords} containing
x-coordinates. Default = NULL, which selects the first column.}

\item{y_coord}{(character) Name of column in \code{spatialCoords} containing
y-coordinates. Default = NULL, which selects the second column.}

\item{sample_id}{(character) Name of column in \code{colData} containing
sample IDs. For datasets with multiple samples, this is used to plot
multiple panels (one per sample) using facetting. Default = NULL.}

\item{in_tissue}{(character) Name of column in \code{colData} identifying
spots over tissue, e.g. "in_tissue" for 10x Genomics Visium data. If this
argument is provided, only spots over tissue will be shown. Alternatively,
set to NULL to display all spots. Default = \code{"in_tissue"}.}

\item{annotate}{(character) Name of column in \code{colData} containing
values to annotate spots with colors, e.g. cluster labels (discrete values)
or total UMI counts (continuous values).}

\item{palette}{(character) Color palette for annotation. Options for discrete
labels (e.g. cluster labels) are "libd_layer_colors", "Okabe-Ito", or a
vector of color names or hex values. For continuous values (e.g. total UMI
counts), provide a vector of length 2 for the low and high range, e.g.
\code{c("gray90", "navy")}. Default = \code{"libd_layer_colors"}.}

\item{trans}{To do. Default = \code{"identity"}.}

\item{assay}{To do. Default = \code{"counts"}.}

\item{legend_position}{To do. Default = \code{"right"}.}

\item{point_size}{(numeric) Point size for \code{geom_point()}. Default =
0.3.}

\item{show_axis}{To do. Default = FALSE.}

\item{text_by}{(character) Column name of the annotation to apply on top of
each cluster. Usually should put it the same as `annotate = `. unless you
have another intended `text_by` column, e.g. with more readable classes or
shorter strings. Only used for categorical `annotate = `. Default = NULL.}

\item{text_by_size}{(numerical) Text size hovering over each cluster. Default
= 5.}

\item{text_by_color}{(character) Color string or hex code. Default =
\code{"black"}.}

\item{y_reverse}{(logical) Whether to reverse y coordinates, which is often
required for 10x Genomics Visium data. Default = TRUE.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, labels, formatting, etc).
}
\description{
Plotting functions for spatially resolved transcriptomics data.
}
\details{
Function to plot spot-based spatially resolved transcriptomics data stored in
a \code{SpatialExperiment} object.

This function generates a plot in spatial coordinates (e.g. x-y coordinates
on a tissue slide), along with annotation such as cluster labels or total UMI
counts.
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
plotSpots(spe, annotate = "ground_truth")

}
\author{
Lukas M. Weber with modifications by Yixing E. Dong
}
