% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVisium.R
\name{plotVisium}
\alias{plotVisium}
\title{plotVisium}
\usage{
plotVisium(
  spe,
  spots = TRUE,
  fill = NULL,
  highlight = NULL,
  facets = "sample_id",
  image = TRUE,
  assay = "logcounts",
  trans = "identity",
  x_coord = "x",
  y_coord = "y",
  y_reverse = FALSE,
  sample_ids = NULL,
  image_ids = NULL,
  palette = NULL
)
}
\arguments{
\item{spe}{(SpatialExperiment) Input data object.}

\item{spots}{(logical) Whether to display spots (spatial barcodes) as points.
Default = TRUE.}

\item{fill}{(character) Column in \code{colData} to use to fill points by
color. If \code{fill} contains a numeric column (e.g. total UMI counts), a
continuous color scale will be used. If \code{fill} contains a factor (e.g.
cluster labels), a discrete color scale will be used. Default = NULL.}

\item{highlight}{(character) Column in \code{spatialData} to use to highlight
points by outlining them. For example, \code{in_tissue} will highlight
spots overlapping with tissue. Default = \code{in_tissue}. Set to NULL to
disable.}

\item{facets}{(character) Column in \code{colData} to use to facet plots,
i.e. show multiple panels of plots. Default = "sample_id". Set to NULL to
disable.}

\item{image}{(logical) Whether to show histology image as background. Default
= TRUE.}

\item{assay}{(character) Name of assay data to use when \code{fill} is in
\code{rownames(spe)}. Should be one of \code{assayNames(spe)}.}

\item{trans}{Transformation to apply for continuous scales. Ignored unless
\code{fill} is numeric, e.g. feature expression. (See
\code{\link{ggplot2}{continuous_scale}} for valid options.)}

\item{x_coord}{(character) Column in \code{spatialCoords} containing
x-coordinates. Default = "x".}

\item{y_coord}{(character) Column in \code{spatialCoords} containing
y-coordinates. Default = "y".}

\item{y_reverse}{(logical) Whether to reverse y coordinates, which may be
required for Visium data, depending on the orientation of the raw data.
Default = FALSE.}

\item{sample_ids}{(character) Samples to show, if multiple samples are
available. Default = NULL (show all samples).}

\item{image_ids}{(character) Images to show, if multiple images are
available. Default = NULL (show all images).}

\item{palette}{(character) Color palette for points. Options for discrete
labels are "libd_layer_colors", "Okabe-Ito", or a custom vector of hex
color codes. Options for continuous values are "viridis", a single color
name (e.g. "red", "navy", etc), or a vector of length two containing color
names for each end of the scale. Default = "libd_layer_colors" for discrete
data, and "viridis" for continuous data.}
}
\value{
Returns a ggplot object. Additional plot elements can be added as
  ggplot elements (e.g. title, customized formatting, etc).
}
\description{
Plots for spatially resolved transcriptomics data from the 10x Genomics
Visium platform
}
\details{
Function to generate plots for spatially resolved transcriptomics datasets
from the 10x Genomics Visium spatially platform.

This function generates a plot for spot-based spatially resolved
transcriptomics data from the 10x Genomics Visium platform, with several
options available to adjust the plot type and style.
}
\examples{
library(ggspavis)
library(STexampleData)

spe <- load_data("Visium_mouseCoronal")

# color by x coordinate, highlight in-tissue spots
plotVisium(spe, fill = "x", highlight = "in_tissue")

# subset in-tissue spots
sub <- spe[, as.logical(spatialData(spe)$in_tissue)]

# color by feature counts, don't include image
rownames(sub) <- make.names(rowData(sub)$gene_name)
plotVisium(sub, fill = "Gad2", assay = "counts", image = FALSE)

}
\author{
Helena L. Crowell with modifications by Lukas M. Weber
}
